TTA sigmoid (#938)
In the current implementation, the network infers background to have
negative values, whereas during the inverse, the padded regions have
values of 0. This causes a +ve standard deviation in regions that are
sometimes inferred and other times are padded back in (i.e., just
outside the brain). Using a sigmoid avoids this as it puts the inferred
background close to zero, too.

# Current


![image](https://user-images.githubusercontent.com/33289025/190447574-9fc6e13e-885e-4296-99a7-0b83f449443e.png)

# Proposed


![image](https://user-images.githubusercontent.com/33289025/190447643-05ba3fff-18b9-41b4-8bd6-9e48a24f595d.png)

### Checks
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Notebook runs automatically `./runner [-p <regex_pattern>]`