change datasets from task 09 to 04 for all auto3dseg notebooks (#997)
Signed-off-by: Mingxin Zheng
<18563433+mingxin-zheng@users.noreply.github.com>

Fixes #993 , 
Fixes #992 , 
Fixes #972 .

### Description
With the recent updates in MONAI core and research contributions, we are
able to fit all models for smaller datasets. Therefore, I suggest
changing all the datasets from the MSD task 09 to task 04, which has
much smaller size and easier to run on different machines. Notebooks are
tested on two RTX 3080Ti with 12GB of mem in each.


### Checks
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [ ] Notebook runs automatically `./runner [-p <regex_pattern>]`

Signed-off-by: Mingxin Zheng <18563433+mingxin-zheng@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>