Update the CAI tutorials with the newest bundle (#1319)
### Description
Currently, this CAI bundle specified the downloaded bundle version in
the tutorials. This makes it difficult to maintain the tutorial, since
the bundles are constantly updated. Therefore, the bundle download code
in the tutorial should track the newest bundles and remove the version
parameter in it.

### Checks
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [ ] Avoid including large-size files in the PR.
- [ ] Clean up long text outputs from code cells in the notebook.
- [ ] For security purposes, please check the contents and remove any
sensitive info such as user names and private key.
- [ ] Ensure (1) hyperlinks and markdown anchors are working (2) use
relative paths for tutorial repo files (3) put figure and graphs in the
`./figure` folder
- [ ] Notebook runs automatically `./runner.sh -t <path to .ipynb file>`

Signed-off-by: binliu <binliu@nvidia.com>