Increase duration in nsys profile  (#1906)
Fixes #1905 

Increase the profiling duration in nsys, as the KeyError seems to occur
because the profile did not have sufficient time to reach the
post-processing step.

### Checks
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [ ] Avoid including large-size files in the PR.
- [ ] Clean up long text outputs from code cells in the notebook.
- [ ] For security purposes, please check the contents and remove any
sensitive info such as user names and private key.
- [ ] Ensure (1) hyperlinks and markdown anchors are working (2) use
relative paths for tutorial repo files (3) put figure and graphs in the
`./figure` folder
- [ ] Notebook runs automatically `./runner.sh -t <path to .ipynb file>`

Signed-off-by: YunLiu <55491388+KumoLiu@users.noreply.github.com>